backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
yagmail
scikit-learn
pydantic
python-multipart
keyring